update package list if a package isn't installed
Currently, the package list is only refreshed if it doesn't
exist. However, since packages seem to be regularly deleted on MELPA
when new versions come out (which is often), we want to update our
package list if we have to install any package. This still saves us from
having to update our list on every start, so in most cases, startup time
should be fast enough.